SummarizePrivateDoccument/requirements.txt
langchain
langgraph
langchain-community
langchain-huggingface
langchain-openai
sentence-transformers
faiss-cpu
torch
tiktoken
python-dotenv
pypdf
